completed refactoring of process_base_dataset.py
